[docs] Extract talks section into transforms module from llms.txt script
